feat(dashboard): implement transfer transaction mutation and refactor handling logic
