+++ apps/wallet-dashboard/hooks/index.ts
@@ -11,3 +11,4 @@ export * from './useGetCurrentEpochStartTimestamp';
 export * from './useTimelockedUnstakeTransaction';
 export * from './useExplorerLinkGetter';
 export * from './useGetStardustMigratableObjects';
+export * from './useTransferTransaction.ts';
